# api/service/http/docs/swagger.yaml
components:
  schemas:
    service.attachment:
      properties:
        attachmentId:
          type: string
        displayName:
          type: string
        documentType:
          type: string
        file:
          $ref: '#/components/schemas/service.file'
        issuedDate:
          type: string
      type: object
    service.file:
      properties:
        uri:
          type: string
      type: object
    service.listResponse:
      properties:
        attachments:
          items:
            $ref: '#/components/schemas/service.attachment'
          type: array
          uniqueItems: false
        vehicleVin:
          type: string
      type: object
externalDocs:
  description: ""
  url: ""
info:
  description: Deliberately dissimilar Service upstream (A5). Nested camelCase, RFC3339
    dates.
  title: Service System mock
  version: 0.1.0
openapi: 3.1.0
paths:
  /healthz:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
          description: OK
        "503":
          content:
            application/json:
              schema:
                type: string
          description: Service Unavailable
      summary: Liveness probe
  /service/v1/vehicles/{vin}/attachments:
    get:
      parameters:
      - description: Vehicle identification number
        in: path
        name: vin
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/service.listResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                type: string
          description: Service Unavailable
      summary: List service attachments for a VIN
servers:
- url: /