# internal/service/modules/service/handler.go
package service

import (
	"log/slog"
	"net/http"
	"strings"

	"github.com/ducnd58233/unified-document-viewer/internal/shared/infra/httpserver"
	"github.com/ducnd58233/unified-document-viewer/internal/shared/mockfault"
	"github.com/ducnd58233/unified-document-viewer/internal/shared/mockseed"
	"github.com/ducnd58233/unified-document-viewer/internal/shared/observability"
)

// Options wires fault injection, file.uri base URL, and optional random attachments.
type Options struct {
	Fault    mockfault.Config
	BaseURL  string
	Generate bool
	IntN     func(n int) int
	Log      observability.Logger
}

// New serves GET /healthz and GET /service/v1/vehicles/{vin}/attachments.
func New(opt Options) http.Handler {
	base := strings.TrimRight(opt.BaseURL, "/")
	if base == "" {
		base = defaultBaseURL
	}
	s := server{
		fault:    opt.Fault,
		docs:     catalog(base),
		baseURL:  base,
		generate: opt.Generate,
		intN:     opt.IntN,
		log:      opt.Log,
	}
	mux := http.NewServeMux()
	mux.HandleFunc("GET /healthz", s.healthz)
	mux.HandleFunc("GET /service/v1/vehicles/{vin}/attachments", s.list)
	return mux
}

type server struct {
	fault    mockfault.Config
	docs     map[string][]attachment
	baseURL  string
	generate bool
	intN     func(n int) int
	log      observability.Logger
}

// healthz godoc
// @Summary Liveness probe
// @Success 200 {string} string
// @Failure 503 {string} string
// @Router /healthz [get]
func (s server) healthz(w http.ResponseWriter, _ *http.Request) {
	if s.fault.Down {
		w.WriteHeader(http.StatusServiceUnavailable)
		return
	}
	w.WriteHeader(http.StatusOK)
}

// list godoc
// @Summary List service attachments for a VIN
// @Param vin path string true "Vehicle identification number"
// @Success 200 {object} listResponse
// @Failure 500 {string} string
// @Failure 503 {string} string
// @Router /service/v1/vehicles/{vin}/attachments [get]
func (s server) list(w http.ResponseWriter, r *http.Request) {
	vin := r.PathValue("vin")
	fault := s.fault.Apply(r.Context(), w)
	if fault.Stop {
		s.traceFault(r, vin, fault)
		return
	}
	atts, ok := s.docs[vin]
	generated := false
	if !ok {
		if s.generate {
			out, err := generateAttachments(vin, s.baseURL, s.intN)
			if err != nil {
				s.traceError(r, vin, "generate", err)
				w.WriteHeader(http.StatusInternalServerError)
				return
			}
			atts = out
			generated = true
		} else {
			atts = []attachment{}
		}
	}
	s.traceOK(r, vin, fault, len(atts), generated)
	httpserver.JSON(w, http.StatusOK, listResponse{VehicleVIN: vin, Attachments: atts})
}

func (s server) baseAttrs(r *http.Request, vin string, fault mockfault.Result) []any {
	return []any{
		slog.String("request_id", httpserver.RequestIDFrom(r.Context())),
		slog.String("method", r.Method),
		slog.String("route", "/service/v1/vehicles/{vin}/attachments"),
		slog.String("vin_suffix", mockseed.Suffix(vin)),
		slog.String("fault", string(fault.Kind)),
		slog.Duration("latency", fault.Latency),
	}
}

func (s server) traceOK(r *http.Request, vin string, fault mockfault.Result, attachments int, generated bool) {
	if s.log == nil {
		return
	}
	s.log.Info("service response", append(s.baseAttrs(r, vin, fault),
		slog.Int("attachments", attachments),
		slog.Bool("generated", generated),
		slog.Int("status", http.StatusOK),
	)...)
}

func (s server) traceFault(r *http.Request, vin string, fault mockfault.Result) {
	if s.log == nil {
		return
	}
	status := 0
	switch fault.Kind {
	case mockfault.KindDown:
		status = http.StatusServiceUnavailable
	case mockfault.KindError:
		status = http.StatusInternalServerError
	}
	s.log.Warn("service fault", append(s.baseAttrs(r, vin, fault), slog.Int("status", status))...)
}

func (s server) traceError(r *http.Request, vin, op string, err error) {
	if s.log == nil {
		return
	}
	s.log.Error("service error",
		slog.String("request_id", httpserver.RequestIDFrom(r.Context())),
		slog.String("vin_suffix", mockseed.Suffix(vin)),
		slog.String("op", op),
		slog.String("error", err.Error()),
	)
}